TransvisionFilter: sort itineraries by duration/taxi
When selecting the least-taxi/least-transfers itineraries are grouped
into bucket. Since the itineraries weren't sorted in the buckets, the
selected itinerary wasn't the one with the least duration.

By sorting the itineraries, the selected itinerary is deterministic, and
the removing similar ones may depend on the one with least duration
being used to filter.